Updated 07/04

+++ src/main/java/com/fullstack/Backend/entities/Ram.java
@@ -19,5 +19,5 @@ import lombok.Setter;
 @Table(name = "Rams")
 public class Ram extends BaseEntity{
 	@Column(nullable = false)
-	private int size;
+	private Long size;
 }
